+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [Unreleased]
+
+### Fixed
+
+- `prov:generatedBy` in output graphs now refers to a plugin IRI instead of a literal
+
+### Changed
+
+- Keep original output ("No explanations found.") if no inconsistencies found with Validate plugin
+
+
 ## [1.0.0beta1] 2024-07-01
 
 ### Fixed
@@ -13,14 +24,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Changed
 
 - complete validation for IRI parameters
-- Remove "Annnotate inferred subclass axioms" parameter
+- remove "Annnotate inferred subclass axioms" parameter in Reason plugin
 
 ## [1.0.0alpha3] 2024-06-28
 
 ### Added
 
 - "Annotate inferred axioms" parameter in Reason plugin
-- "Maximum RAM percentage" parameter in Reason and Validate plugins
+- "Maximum RAM percentage" parameter
 
 ### Changed